objects: rename ai/ to creatures/
