+++ meson.build
@@ -133,32 +133,32 @@ sources = [
   'src/game/los.c',
   'src/game/lot.c',
   'src/game/music.c',
-  'src/game/objects/ai/ape.c',
-  'src/game/objects/ai/bacon_lara.c',
-  'src/game/objects/ai/baldy.c',
-  'src/game/objects/ai/bat.c',
-  'src/game/objects/ai/bear.c',
-  'src/game/objects/ai/centaur.c',
-  'src/game/objects/ai/cowboy.c',
-  'src/game/objects/ai/crocodile.c',
-  'src/game/objects/ai/larson.c',
-  'src/game/objects/ai/lion.c',
-  'src/game/objects/ai/mummy.c',
-  'src/game/objects/ai/mutant.c',
-  'src/game/objects/ai/natla.c',
-  'src/game/objects/ai/pierre.c',
-  'src/game/objects/ai/pod.c',
-  'src/game/objects/ai/raptor.c',
-  'src/game/objects/ai/rat.c',
-  'src/game/objects/ai/skate_kid.c',
-  'src/game/objects/ai/statue.c',
-  'src/game/objects/ai/torso.c',
-  'src/game/objects/ai/trex.c',
-  'src/game/objects/ai/wolf.c',
   'src/game/objects/boat.c',
   'src/game/objects/bridge.c',
   'src/game/objects/cabin.c',
   'src/game/objects/cog.c',
+  'src/game/objects/creatures/ape.c',
+  'src/game/objects/creatures/bacon_lara.c',
+  'src/game/objects/creatures/baldy.c',
+  'src/game/objects/creatures/bat.c',
+  'src/game/objects/creatures/bear.c',
+  'src/game/objects/creatures/centaur.c',
+  'src/game/objects/creatures/cowboy.c',
+  'src/game/objects/creatures/crocodile.c',
+  'src/game/objects/creatures/larson.c',
+  'src/game/objects/creatures/lion.c',
+  'src/game/objects/creatures/mummy.c',
+  'src/game/objects/creatures/mutant.c',
+  'src/game/objects/creatures/natla.c',
+  'src/game/objects/creatures/pierre.c',
+  'src/game/objects/creatures/pod.c',
+  'src/game/objects/creatures/raptor.c',
+  'src/game/objects/creatures/rat.c',
+  'src/game/objects/creatures/skate_kid.c',
+  'src/game/objects/creatures/statue.c',
+  'src/game/objects/creatures/torso.c',
+  'src/game/objects/creatures/trex.c',
+  'src/game/objects/creatures/wolf.c',
   'src/game/objects/door.c',
   'src/game/objects/effects/blood.c',
   'src/game/objects/effects/body_part.c',
