+++ src/game/objects/creatures/ape.h
