+++ src/game/objects/creatures/bacon_lara.h
