+++ src/game/objects/creatures/baldy.h
