+++ src/game/objects/creatures/bat.h
