+++ src/game/objects/creatures/bear.h
