+++ src/game/objects/creatures/cowboy.h
