+++ src/game/objects/creatures/crocodile.h
