+++ src/game/objects/creatures/larson.h
