+++ src/game/objects/creatures/lion.h
