+++ src/game/objects/creatures/mummy.h
