+++ src/game/objects/creatures/mutant.h
