+++ src/game/objects/creatures/natla.h
