+++ src/game/objects/creatures/pierre.h
