+++ src/game/objects/creatures/rat.h
