+++ src/game/objects/creatures/skate_kid.h
