+++ src/game/objects/creatures/statue.h
