+++ src/game/objects/creatures/torso.h
