+++ src/game/objects/creatures/trex.h
