+++ src/game/objects/creatures/wolf.h
