+++ src/game/setup.c
@@ -9,32 +9,32 @@
 #include "game/inv.h"
 #include "game/lara.h"
 #include "game/level.h"
-#include "game/objects/ai/ape.h"
-#include "game/objects/ai/bacon_lara.h"
-#include "game/objects/ai/baldy.h"
-#include "game/objects/ai/bat.h"
-#include "game/objects/ai/bear.h"
-#include "game/objects/ai/centaur.h"
-#include "game/objects/ai/cowboy.h"
-#include "game/objects/ai/crocodile.h"
-#include "game/objects/ai/larson.h"
-#include "game/objects/ai/lion.h"
-#include "game/objects/ai/mummy.h"
-#include "game/objects/ai/mutant.h"
-#include "game/objects/ai/natla.h"
-#include "game/objects/ai/pierre.h"
-#include "game/objects/ai/pod.h"
-#include "game/objects/ai/raptor.h"
-#include "game/objects/ai/rat.h"
-#include "game/objects/ai/skate_kid.h"
-#include "game/objects/ai/statue.h"
-#include "game/objects/ai/torso.h"
-#include "game/objects/ai/trex.h"
-#include "game/objects/ai/wolf.h"
 #include "game/objects/boat.h"
 #include "game/objects/bridge.h"
 #include "game/objects/cabin.h"
 #include "game/objects/cog.h"
+#include "game/objects/creatures/ape.h"
+#include "game/objects/creatures/bacon_lara.h"
+#include "game/objects/creatures/baldy.h"
+#include "game/objects/creatures/bat.h"
+#include "game/objects/creatures/bear.h"
+#include "game/objects/creatures/centaur.h"
+#include "game/objects/creatures/cowboy.h"
+#include "game/objects/creatures/crocodile.h"
+#include "game/objects/creatures/larson.h"
+#include "game/objects/creatures/lion.h"
+#include "game/objects/creatures/mummy.h"
+#include "game/objects/creatures/mutant.h"
+#include "game/objects/creatures/natla.h"
+#include "game/objects/creatures/pierre.h"
+#include "game/objects/creatures/pod.h"
+#include "game/objects/creatures/raptor.h"
+#include "game/objects/creatures/rat.h"
+#include "game/objects/creatures/skate_kid.h"
+#include "game/objects/creatures/statue.h"
+#include "game/objects/creatures/torso.h"
+#include "game/objects/creatures/trex.h"
+#include "game/objects/creatures/wolf.h"
 #include "game/objects/door.h"
 #include "game/objects/effects/blood.h"
 #include "game/objects/effects/body_part.h"
